requirements.txt
super-gradients
flask-socketio
fire
fastcore
